requirements.txt
Django
Pandas
azure-storage-blob 
azure-identity
django-environ